Initial Commit
This commit includes much of the code that will exist in the final version of the assignment. The code for the encoding represantations for Binary, Base64, ASCII and padding are finalized. The code for compression algorithms such as Huffman encoding, LempelZiv78 and Arithmetic Encoding(this implementation is bad) is also finalized, but the arithmetic encoding implementation is pretty bad.. For error correcting codes the the Hamming74 implementation is also finalized and the same applies to the CRC implementation. What is left is better look at the retransmission mechanism, the addition of some metrics, implementation of the API endpoints and the finalization of the UI.