Added The Rest Of The Project
I added the remaining code of the project. This includes the finalized UI with the appropriate user interactivity. The Shannon Fano compression algorithm was also added to the available loseless compression algorithms(other 2 being Huffman compression and LempelZiv78). I decided to add it, because of its simplicity and its similarity with the Huffman compression. Another thing that was added were statistics such as entropy, errors added and errors detected and retrasmission count. When it comes to error insertion I made sure that error insertion works without any bugs with all the compression algorithms and the error correction algorithms. As for the error correcting algorithms I implemented a very basic emulation of a retransmission mechanism. That specific implementation breaks the RESTful principles, but I think that it is ok for the assignment. Finally I greatly refactored the c ode, so that it is now more readable and modular. I must admit that I did not add in the compression codes Arithmetic encoding since I did not manage to make the algorithm work, but the implementation of the algorithm still exists in the library and might be added in a later update(probably unlikely).